refactor: Phase 1 Type Dehydration - Move BuildPartitionForDocument to ids package (#47)
- Extract BuildPartitionForDocument from types/partition.go to ids/partition_builder.go
- Move corresponding test to ids/partition_builder_test.go
- Update caller in id_generator.go to use local function
- Maintain all existing functionality with passing tests
