+++ types/partition.go
@@ -130,42 +130,3 @@ func (pm PartitionMap) Get(partition Partition) []Document {
 func (pm PartitionMap) Count(partition Partition) int {
 	return len(pm[partition.Key()])
 }
-
-// BuildPartitionForDocument creates a Partition from a document's dimensions
-func BuildPartitionForDocument(doc Document, dimensionSet *DimensionSet) Partition {
-	var values []DimensionValue
-
-	// Build dimension values in order
-	for _, dim := range dimensionSet.All() {
-		value := ""
-
-		switch dim.Type {
-		case Enumerated:
-			// Get value from dimensions map
-			if v, exists := doc.Dimensions[dim.Name]; exists {
-				value = fmt.Sprintf("%v", v)
-			} else {
-				// Use default value
-				value = dim.DefaultValue
-			}
-
-		case Hierarchical:
-			// For hierarchical dimensions, use the parent reference
-			if v, exists := doc.Dimensions[dim.RefField]; exists {
-				value = fmt.Sprintf("%v", v)
-			}
-		}
-
-		if value != "" {
-			values = append(values, DimensionValue{
-				Dimension: dim.Name,
-				Value:     value,
-			})
-		}
-	}
-
-	return Partition{
-		Values:   values,
-		Position: 0, // Position will be set during ID generation
-	}
-}
